Sleeping!
Not kidding this time, sleeping at an inn will refill your health completely (less helpful early on, probably will be in later levels). Inn Cost is random and changes after each encounter (when it's an option), and goes up with level. It won't change just because you back out and come back in, you gotta work for a better price.