No more crashing because of typos!
Removed all instances of the Unidentified Exception, although I'm keeping it around just in case. Now everything you could do before to make the game crash will just loop around and make you try again. The only time the game will actually quit is if the Distributor tries to fetch a bad ID.
May eventually use the Unidentified exception again, but the process of containing things that throw exceptions is so excruciatingly painful that I think it'll be a relic from here on out.